submodulize
