+++ src/SensitiveFloats.jl
@@ -1,5 +1,15 @@
+module SensitiveFloats
 
-export multiply_sfs!, add_scaled_sfs!, combine_sfs!, add_sources_sf!
+export SensitiveFloat,
+       zero_sensitive_float,
+       clear!,
+       multiply_sfs!,
+       add_scaled_sfs!,
+       combine_sfs!,
+       add_sources_sf!,
+       set_hess!
+
+import ..Types: ParamSet, getids
 
 """
 A function value and its derivative with respect to its arguments.
@@ -15,7 +25,7 @@ Attributes:
   hs: An array of per-source Hessians.  This will generally be reserved
       for the Hessian of brightness values that depend only on one source.
 """
-type SensitiveFloat{ParamType <: CelesteTypes.ParamSet, NumType <: Number}
+type SensitiveFloat{ParamType <: ParamSet, NumType <: Number}
     # Actually a single value, but an Array to avoid memory allocation
     v::Vector{NumType}
 
@@ -34,7 +44,7 @@ end
 """
 Set a SensitiveFloat's hessian term, maintaining symmetry.
 """
-function set_hess!{ParamType <: CelesteTypes.ParamSet, NumType <: Number}(
+function set_hess!{ParamType <: ParamSet, NumType <: Number}(
     sf::SensitiveFloat{ParamType, NumType},
     i::Int, j::Int, v::NumType)
   i != j ?
@@ -44,7 +54,7 @@ function set_hess!{ParamType <: CelesteTypes.ParamSet, NumType <: Number}(
     true # Set definite return type
 end
 
-function zero_sensitive_float{ParamType <: CelesteTypes.ParamSet}(
+function zero_sensitive_float{ParamType <: ParamSet}(
   ::Type{ParamType}, NumType::DataType, local_S::Int)
     local_P = length(ParamType)
 
@@ -55,7 +65,7 @@ function zero_sensitive_float{ParamType <: CelesteTypes.ParamSet}(
 end
 
 
-function zero_sensitive_float{ParamType <: CelesteTypes.ParamSet}(
+function zero_sensitive_float{ParamType <: ParamSet}(
   ::Type{ParamType}, NumType::DataType)
     # Default to a single source.
     zero_sensitive_float(ParamType, NumType, 1)
@@ -63,26 +73,26 @@ end
 
 
 # If no type is specified, default to using Float64.
-function zero_sensitive_float{ParamType <: CelesteTypes.ParamSet}(
+function zero_sensitive_float{ParamType <: ParamSet}(
   param_arg::Type{ParamType}, local_S::Int)
     zero_sensitive_float(param_arg, Float64, local_S)
 end
 
 
-function zero_sensitive_float{ParamType <: CelesteTypes.ParamSet}(
+function zero_sensitive_float{ParamType <: ParamSet}(
   param_arg::Type{ParamType})
     zero_sensitive_float(param_arg, Float64, 1)
 end
 
 
-function clear!{ParamType <: CelesteTypes.ParamSet, NumType <: Number}(
+function clear!{ParamType <: ParamSet, NumType <: Number}(
   sp::SensitiveFloat{ParamType, NumType})
 
   clear!(sp, true)
 end
 
 
-function clear!{ParamType <: CelesteTypes.ParamSet, NumType <: Number}(
+function clear!{ParamType <: ParamSet, NumType <: Number}(
   sp::SensitiveFloat{ParamType, NumType}, clear_hessian::Bool)
 
     fill!(sp.v, zero(NumType))
@@ -98,7 +108,7 @@ end
 """
 Factor out the hessian part of combine_sfs!.
 """
-function combine_sfs_hessian!{ParamType <: CelesteTypes.ParamSet,
+function combine_sfs_hessian!{ParamType <: ParamSet,
                               T1 <: Number, T2 <: Number, T3 <: Number}(
     sf1::SensitiveFloat{ParamType, T1},
     sf2::SensitiveFloat{ParamType, T1},
@@ -135,7 +145,7 @@ each evaluated at (sf1, sf2).
 The result is stored in sf_result.  The order is done in such a way that
 it can overwrite sf1 or sf2 and still be accurate.
 """
-function combine_sfs!{ParamType <: CelesteTypes.ParamSet,
+function combine_sfs!{ParamType <: ParamSet,
                       T1 <: Number, T2 <: Number, T3 <: Number}(
     sf1::SensitiveFloat{ParamType, T1},
     sf2::SensitiveFloat{ParamType, T1},
@@ -173,7 +183,7 @@ each evaluated at (sf1, sf2).
 
 The result is stored in sf1.
 """
-function combine_sfs!{ParamType <: CelesteTypes.ParamSet,
+function combine_sfs!{ParamType <: ParamSet,
                       T1 <: Number, T2 <: Number, T3 <: Number}(
     sf1::SensitiveFloat{ParamType, T1},
     sf2::SensitiveFloat{ParamType, T1},
@@ -191,7 +201,7 @@ const multiply_sfs_hess = Float64[0 1; 1 0]
 """
 TODO: don't ignore the ids arguments.
 """
-function multiply_sfs!{ParamType <: CelesteTypes.ParamSet, NumType <: Number}(
+function multiply_sfs!{ParamType <: ParamSet, NumType <: Number}(
     sf1::SensitiveFloat{ParamType, NumType},
     sf2::SensitiveFloat{ParamType, NumType};
     ids1::Vector{Int}=collect(1:length(ParamType)),
@@ -211,7 +221,7 @@ end
 """
 Update sf1 in place with (sf1 + scale * sf2).
 """
-function add_scaled_sfs!{ParamType <: CelesteTypes.ParamSet, NumType <: Number}(
+function add_scaled_sfs!{ParamType <: ParamSet, NumType <: Number}(
     sf1::SensitiveFloat{ParamType, NumType},
     sf2::SensitiveFloat{ParamType, NumType},
     scale::Float64, calculate_hessian::Bool)
@@ -239,7 +249,7 @@ end
 Adds sf2_s to sf1, where sf1 is sensitive to multiple sources and sf2_s is only
 sensitive to source s.
 """
-function add_sources_sf!{ParamType <: CelesteTypes.ParamSet, NumType <: Number}(
+function add_sources_sf!{ParamType <: ParamSet, NumType <: Number}(
     sf_all::SensitiveFloat{ParamType, NumType},
     sf_s::SensitiveFloat{ParamType, NumType},
     s::Int, calculate_hessian::Bool)
@@ -272,4 +282,6 @@ function add_sources_sf!{ParamType <: CelesteTypes.ParamSet, NumType <: Number}(
   end
 
   true # Set definite return type
+end
+
 end
