+++ src/source_brightness.jl
